requirements.txt
Django<=4.0.3
djangorestframework<=3.13.1
psycopg2>=2.9.3
drf-nested-routers==0.93.4
Faker==13.3.0

flake8<=4.0.1
